backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
psycopg2-binary
alembic
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
